requirements/dev.txt
-r base.txt
sphinx
mysql-python
django-debug-toolbar
